add dask_array_ops.py and fixes for dask/rolling
