+++ xarray/core/ops.py
@@ -28,11 +28,6 @@ except ImportError:
     bn = np
     has_bottleneck = False
 
-try:
-    import dask.array as da
-except ImportError:
-    pass
-
 
 UNARY_OPS = ['neg', 'pos', 'abs', 'invert']
 CMP_BINARY_OPS = ['lt', 'le', 'ge', 'gt']
@@ -247,41 +242,6 @@ def rolling_count(rolling):
     enough_periods = rolling_count >= rolling.min_periods
 
     return rolling_count.where(enough_periods)
-
-
-def dask_rolling_wrapper(moving_func, a, window, min_count=None, axis=-1):
-    '''wrapper to apply bottleneck moving window funcs on dask arrays'''
-    # inputs for ghost
-    if axis < 0:
-        axis = a.ndim + axis
-    depth = {d: 0 for d in range(a.ndim)}
-    depth[axis] = window - 1
-    boundary = {d: np.nan for d in range(a.ndim)}
-    # create ghosted arrays
-    ag = da.ghost.ghost(a, depth=depth, boundary=boundary)
-    # apply rolling func
-    out = ag.map_blocks(moving_func, window, min_count=min_count,
-                        axis=axis, dtype=a.dtype)
-    # trim array
-    result = da.ghost.trim_internal(out, depth)
-    return result
-
-
-def dask_rolling_wrapper_without_min_count(moving_func, a, window, axis=-1):
-    '''wrapper to apply bottleneck moving window funcs on dask arrays'''
-    # inputs for ghost
-    if axis < 0:
-        axis = a.ndim + axis
-    depth = {d: 0 for d in range(a.ndim)}
-    depth[axis] = window - 1
-    boundary = {d: np.nan for d in range(a.ndim)}
-    # create ghosted arrays
-    ag = da.ghost.ghost(a, depth=depth, boundary=boundary)
-    # apply rolling func
-    out = ag.map_blocks(moving_func, window, axis=axis, dtype=a.dtype)
-    # trim array
-    result = da.ghost.trim_internal(out, depth)
-    return result
 
 
 def inject_reduce_methods(cls):
